fix: imposto valor float

+++ src/app/Controller/Api/Imposto/ImpostoController.php
@@ -296,7 +296,7 @@ class ImpostoController extends Controller
     {
         $body = $this->request->body() ?? [];
         $atributos = [
-            'valor' => $body['valor'] ?? null,
+            'valor' => (float) ($body['valor'] ?? null),
         ];
 
         try {
